Add dual-port ram.
Tried true dual-port. Both single-port and dual-port has the same
issue that they need address ready before the clk posedge. But in
the one cycle design the address was generated during the cycle.

Reset signal is set to #22, a little more exceeds the clk rising edge.
This can keep the PC stay the same for this cycle. Because even
during the rst, part of the cpu is functionning. It gets the next IP
ready.